+++ src/gui/windows/botcheckerwindow.cpp
@@ -116,9 +116,9 @@ public:
         }
 
         const unsigned int curTime = cur_time;
-        const unsigned int sz = mPlayers.size();
+        const size_t sz = mPlayers.size();
         // set up widgets
-        for (unsigned int r = 0; r < sz; ++r)
+        for (size_t r = 0; r < sz; ++r)
         {
             if (!mPlayers.at(r))
                 continue;
